dodge: iOS port — gdextension iOS entries + cross-platform script idioms
- addons/kanama/kanama.gdextension: add ios.debug/release.arm64 entries
  (regenerated by installIosAddon).
- Player.kt: velocity.x/y != 0f -> .toDouble() != 0.0 (Kotlin forbids
  Double/Float == /!=; iOS components are Double, Android/desktop Float).
- SmokeQuit.kt: SceneTree.quit() (static) -> getTree().quit() (instance)
  to match the cross-platform-common idiom; keeps frame-based
  awaitNextFrame loop (now supported on iOS).

Builds + links for iOS device target.

+++ godot-demo-2d-dodge-the-creeps/kotlin-src/Player.kt
@@ -59,13 +59,13 @@ class Player(godotObject: MemorySegment) : KanamaScript<Area2D>(godotObject, ::A
 
         self.position = (self.position + velocity * delta).clamp(Vector2.ZERO, screenSize)
 
-        if (velocity.x != 0f) {
+        if (velocity.x.toDouble() != 0.0) {
             animatedSprite.animation = "right"
             self.rotation = 0.0
             animatedSprite.flipV = false
             trail.rotation = 0.0
             animatedSprite.flipH = velocity.x < 0f
-        } else if (velocity.y != 0f) {
+        } else if (velocity.y.toDouble() != 0.0) {
             animatedSprite.animation = "up"
             self.rotation = if (velocity.y > 0f) Mathf.PI else 0.0
         }
